backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
supabase==2.0.3
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
joblib==1.3.2
python-multipart==0.0.6
PyJWT==2.8.0
